Add statistics entity. Move client stats from Monitor to EventManager
