+++ libafl/src/events/broker_hooks/mod.rs
@@ -15,6 +15,7 @@ use crate::events::llmp::COMPRESS_THRESHOLD;
 use crate::{
     events::{llmp::LLMP_TAG_EVENT_TO_BOTH, BrokerEventResult, Event},
     monitors::Monitor,
+    statistics::manager::ClientStatsManager,
     Error,
 };
 
@@ -37,6 +38,7 @@ pub struct StdLlmpEventHook<I, MT> {
     #[cfg(feature = "llmp_compression")]
     compressor: GzipCompressor,
     phantom: PhantomData<I>,
+    client_stats_manager: ClientStatsManager,
 }
 
 impl<I, MT, SHM, SP> LlmpHook<SHM, SP> for StdLlmpEventHook<I, MT>
@@ -71,7 +73,12 @@ where
                 &*msg
             };
             let event: Event<I> = postcard::from_bytes(event_bytes)?;
-            match Self::handle_in_broker(monitor, client_id, &event)? {
+            match Self::handle_in_broker(
+                monitor,
+                &mut self.client_stats_manager,
+                client_id,
+                &event,
+            )? {
                 BrokerEventResult::Forward => Ok(LlmpMsgHookResult::ForwardToClients),
                 BrokerEventResult::Handled => Ok(LlmpMsgHookResult::Handled),
             }
@@ -81,7 +88,11 @@ where
     }
 
     fn on_timeout(&mut self) -> Result<(), Error> {
-        self.monitor.display("Broker Heartbeat", ClientId(0));
+        self.monitor.display(
+            &mut self.client_stats_manager,
+            "Broker Heartbeat",
+            ClientId(0),
+        );
         Ok(())
     }
 }
@@ -96,6 +107,7 @@ where
             monitor,
             #[cfg(feature = "llmp_compression")]
             compressor: GzipCompressor::with_threshold(COMPRESS_THRESHOLD),
+            client_stats_manager: ClientStatsManager::default(),
             phantom: PhantomData,
         })
     }
@@ -104,6 +116,7 @@ where
     #[expect(clippy::unnecessary_wraps)]
     fn handle_in_broker(
         monitor: &mut MT,
+        client_stats_manager: &mut ClientStatsManager,
         client_id: ClientId,
         event: &Event<I>,
     ) -> Result<BrokerEventResult, Error> {
@@ -119,11 +132,11 @@ where
                     client_id
                 };
 
-                monitor.client_stats_insert(id);
-                monitor.update_client_stats_for(id, |client_stat| {
+                client_stats_manager.client_stats_insert(id);
+                client_stats_manager.update_client_stats_for(id, |client_stat| {
                     client_stat.update_corpus_size(*corpus_size as u64);
                 });
-                monitor.display(event.name(), id);
+                monitor.display(client_stats_manager, event.name(), id);
                 Ok(BrokerEventResult::Forward)
             }
             Event::UpdateExecStats {
@@ -132,24 +145,20 @@ where
                 phantom: _,
             } => {
                 // TODO: The monitor buffer should be added on client add.
-                monitor.client_stats_insert(client_id);
-                monitor.update_client_stats_for(client_id, |client_stat| {
+                client_stats_manager.client_stats_insert(client_id);
+                client_stats_manager.update_client_stats_for(client_id, |client_stat| {
                     client_stat.update_executions(*executions, *time);
                 });
-                monitor.display(event.name(), client_id);
+                monitor.display(client_stats_manager, event.name(), client_id);
                 Ok(BrokerEventResult::Handled)
             }
-            Event::UpdateUserStats {
-                name,
-                value,
-                phantom: _,
-            } => {
-                monitor.client_stats_insert(client_id);
-                monitor.update_client_stats_for(client_id, |client_stat| {
+            Event::UpdateUserStats { name, value, .. } => {
+                client_stats_manager.client_stats_insert(client_id);
+                client_stats_manager.update_client_stats_for(client_id, |client_stat| {
                     client_stat.update_user_stats(name.clone(), value.clone());
                 });
-                monitor.aggregate(name);
-                monitor.display(event.name(), client_id);
+                client_stats_manager.aggregate(name);
+                monitor.display(client_stats_manager, event.name(), client_id);
                 Ok(BrokerEventResult::Handled)
             }
             #[cfg(feature = "introspection")]
@@ -162,8 +171,8 @@ where
                 // TODO: The monitor buffer should be added on client add.
 
                 // Get the client for the staterestorer ID
-                monitor.client_stats_insert(client_id);
-                monitor.update_client_stats_for(client_id, |client_stat| {
+                client_stats_manager.client_stats_insert(client_id);
+                client_stats_manager.update_client_stats_for(client_id, |client_stat| {
                     // Update the normal monitor for this client
                     client_stat.update_executions(*executions, *time);
                     // Update the performance monitor for this client
@@ -171,17 +180,17 @@ where
                 });
 
                 // Display the monitor via `.display` only on core #1
-                monitor.display(event.name(), client_id);
+                monitor.display(client_stats_manager, event.name(), client_id);
 
                 // Correctly handled the event
                 Ok(BrokerEventResult::Handled)
             }
             Event::Objective { objective_size, .. } => {
-                monitor.client_stats_insert(client_id);
-                monitor.update_client_stats_for(client_id, |client_stat| {
+                client_stats_manager.client_stats_insert(client_id);
+                client_stats_manager.update_client_stats_for(client_id, |client_stat| {
                     client_stat.update_objective_size(*objective_size as u64);
                 });
-                monitor.display(event.name(), client_id);
+                monitor.display(client_stats_manager, event.name(), client_id);
                 Ok(BrokerEventResult::Handled)
             }
             Event::Log {
