+++ libafl/src/monitors/disk.rs
@@ -1,6 +1,6 @@
 //! Monitors that wrap a base monitor and also log to disk using different formats like `JSON` and `TOML`.
 
-use alloc::{string::String, vec::Vec};
+use alloc::string::String;
 use core::time::Duration;
 use std::{
     fs::{File, OpenOptions},
@@ -11,7 +11,10 @@ use std::{
 use libafl_bolts::{current_time, format_duration_hms, ClientId};
 use serde_json::json;
 
-use crate::monitors::{ClientStats, Monitor, NopMonitor};
+use crate::{
+    monitors::{Monitor, NopMonitor},
+    statistics::manager::ClientStatsManager,
+};
 
 /// Wrap a monitor and log the current state of the monitor into a Toml file.
 #[derive(Debug, Clone)]
@@ -29,16 +32,6 @@ impl<M> Monitor for OnDiskTomlMonitor<M>
 where
     M: Monitor,
 {
-    /// The client monitor, mutable
-    fn client_stats_mut(&mut self) -> &mut Vec<ClientStats> {
-        self.base.client_stats_mut()
-    }
-
-    /// The client monitor
-    fn client_stats(&self) -> &[ClientStats] {
-        self.base.client_stats()
-    }
-
     /// Time this fuzzing run stated
     fn start_time(&self) -> Duration {
         self.base.start_time()
@@ -49,11 +42,12 @@ where
         self.base.set_start_time(time);
     }
 
-    fn aggregate(&mut self, name: &str) {
-        self.base.aggregate(name);
-    }
-
-    fn display(&mut self, event_msg: &str, sender_id: ClientId) {
+    fn display(
+        &mut self,
+        client_stats_manager: &mut ClientStatsManager,
+        event_msg: &str,
+        sender_id: ClientId,
+    ) {
         let cur_time = current_time();
 
         if cur_time - self.last_update >= self.update_interval {
@@ -73,21 +67,22 @@ executions = {}
 exec_sec = {}
 ",
                 format_duration_hms(&(cur_time - self.start_time())),
-                self.client_stats_count(),
-                self.corpus_size(),
-                self.objective_size(),
-                self.total_execs(),
-                self.execs_per_sec()
+                client_stats_manager.client_stats_count(),
+                client_stats_manager.corpus_size(),
+                client_stats_manager.objective_size(),
+                client_stats_manager.total_execs(),
+                client_stats_manager.execs_per_sec()
             )
             .expect("Failed to write to the Toml file");
 
-            for i in 0..(self.client_stats().len()) {
+            for i in 0..(client_stats_manager.client_stats().len()) {
                 let client_id = ClientId(i as u32);
-                let exec_sec = self.update_client_stats_for(client_id, |client_stat| {
-                    client_stat.execs_per_sec(cur_time)
-                });
+                let exec_sec = client_stats_manager
+                    .update_client_stats_for(client_id, |client_stat| {
+                        client_stat.execs_per_sec(cur_time)
+                    });
 
-                let client = self.client_stats_for(client_id);
+                let client = client_stats_manager.client_stats_for(client_id);
 
                 write!(
                     &mut file,
@@ -102,7 +97,7 @@ exec_sec = {}
                 )
                 .expect("Failed to write to the Toml file");
 
-                for (key, val) in &client.user_monitor {
+                for (key, val) in &client.user_stats {
                     let k: String = key
                         .chars()
                         .map(|c| if c.is_whitespace() { '_' } else { c })
@@ -116,7 +111,8 @@ exec_sec = {}
             drop(file);
         }
 
-        self.base.display(event_msg, sender_id);
+        self.base
+            .display(client_stats_manager, event_msg, sender_id);
     }
 }
 
@@ -197,14 +193,6 @@ where
     F: FnMut(&mut M) -> bool,
     M: Monitor,
 {
-    fn client_stats_mut(&mut self) -> &mut Vec<ClientStats> {
-        self.base.client_stats_mut()
-    }
-
-    fn client_stats(&self) -> &[ClientStats] {
-        self.base.client_stats()
-    }
-
     fn start_time(&self) -> Duration {
         self.base.start_time()
     }
@@ -213,7 +201,12 @@ where
         self.base.set_start_time(time);
     }
 
-    fn display(&mut self, event_msg: &str, sender_id: ClientId) {
+    fn display(
+        &mut self,
+        client_stats_manager: &mut ClientStatsManager,
+        event_msg: &str,
+        sender_id: ClientId,
+    ) {
         if (self.log_record)(&mut self.base) {
             let file = OpenOptions::new()
                 .append(true)
@@ -223,15 +216,16 @@ where
 
             let line = json!({
                 "run_time": current_time() - self.base.start_time(),
-                "clients": self.client_stats_count(),
-                "corpus": self.base.corpus_size(),
-                "objectives": self.base.objective_size(),
-                "executions": self.base.total_execs(),
-                "exec_sec": self.base.execs_per_sec(),
-                "client_stats": self.client_stats(),
+                "clients": client_stats_manager.client_stats_count(),
+                "corpus": client_stats_manager.corpus_size(),
+                "objectives": client_stats_manager.objective_size(),
+                "executions": client_stats_manager.total_execs(),
+                "exec_sec": client_stats_manager.execs_per_sec(),
+                "client_stats": client_stats_manager.client_stats(),
             });
             writeln!(&file, "{line}").expect("Unable to write Json to file");
         }
-        self.base.display(event_msg, sender_id);
+        self.base
+            .display(client_stats_manager, event_msg, sender_id);
     }
 }
