+++ libafl/src/monitors/multi.rs
@@ -8,8 +8,10 @@ use core::{
 
 use libafl_bolts::{current_time, format_duration_hms, ClientId};
 
-use super::Aggregator;
-use crate::monitors::{ClientStats, Monitor};
+use crate::{
+    monitors::{ClientStats, Monitor},
+    statistics::manager::ClientStatsManager,
+};
 
 /// Tracking monitor during fuzzing and display both per-client and cumulative info.
 #[derive(Clone)]
@@ -20,7 +22,6 @@ where
     print_fn: F,
     start_time: Duration,
     client_stats: Vec<ClientStats>,
-    aggregator: Aggregator,
 }
 
 impl<F> Debug for MultiMonitor<F>
@@ -39,16 +40,6 @@ impl<F> Monitor for MultiMonitor<F>
 where
     F: FnMut(&str),
 {
-    /// the client monitor, mutable
-    fn client_stats_mut(&mut self) -> &mut Vec<ClientStats> {
-        &mut self.client_stats
-    }
-
-    /// the client monitor
-    fn client_stats(&self) -> &[ClientStats] {
-        &self.client_stats
-    }
-
     /// Set creation time
     fn set_start_time(&mut self, time: Duration) {
         self.start_time = time;
@@ -59,11 +50,12 @@ where
         self.start_time
     }
 
-    fn aggregate(&mut self, name: &str) {
-        self.aggregator.aggregate(name, &self.client_stats);
-    }
-
-    fn display(&mut self, event_msg: &str, sender_id: ClientId) {
+    fn display(
+        &mut self,
+        client_stats_manager: &mut ClientStatsManager,
+        event_msg: &str,
+        sender_id: ClientId,
+    ) {
         let sender = format!("#{}", sender_id.0);
         let pad = if event_msg.len() + sender.len() < 13 {
             " ".repeat(13 - event_msg.len() - sender.len())
@@ -75,30 +67,30 @@ where
             "[{}]  (GLOBAL) run time: {}, clients: {}, corpus: {}, objectives: {}, executions: {}, exec/sec: {}",
             head,
             format_duration_hms(&(current_time() - self.start_time)),
-            self.client_stats_count(),
-            self.corpus_size(),
-            self.objective_size(),
-            self.total_execs(),
-            self.execs_per_sec_pretty()
+            client_stats_manager.client_stats_count(),
+            client_stats_manager.corpus_size(),
+            client_stats_manager.objective_size(),
+            client_stats_manager.total_execs(),
+            client_stats_manager.execs_per_sec_pretty()
         );
-        for (key, val) in &self.aggregator.aggregated {
+        for (key, val) in client_stats_manager.aggregated() {
             write!(global_fmt, ", {key}: {val}").unwrap();
         }
 
         (self.print_fn)(&global_fmt);
 
-        self.client_stats_insert(sender_id);
+        client_stats_manager.client_stats_insert(sender_id);
         let cur_time = current_time();
-        let exec_sec =
-            self.update_client_stats_for(sender_id, |client| client.execs_per_sec_pretty(cur_time));
-        let client = self.client_stats_for(sender_id);
+        let exec_sec = client_stats_manager
+            .update_client_stats_for(sender_id, |client| client.execs_per_sec_pretty(cur_time));
+        let client = client_stats_manager.client_stats_for(sender_id);
 
         let pad = " ".repeat(head.len());
         let mut fmt = format!(
             " {}   (CLIENT) corpus: {}, objectives: {}, executions: {}, exec/sec: {}",
             pad, client.corpus_size, client.objective_size, client.executions, exec_sec
         );
-        for (key, val) in &client.user_monitor {
+        for (key, val) in &client.user_stats {
             write!(fmt, ", {key}: {val}").unwrap();
         }
         (self.print_fn)(&fmt);
@@ -128,7 +120,6 @@ where
             print_fn,
             start_time: current_time(),
             client_stats: vec![],
-            aggregator: Aggregator::new(),
         }
     }
 
@@ -138,7 +129,6 @@ where
             print_fn,
             start_time,
             client_stats: vec![],
-            aggregator: Aggregator::new(),
         }
     }
 }
