+++ libafl/src/monitors/prometheus.rs
@@ -47,8 +47,8 @@ use prometheus_client::{
 // using tide for the HTTP server library (fast, async, simple)
 use tide::Request;
 
-use super::Aggregator;
-use crate::monitors::{ClientStats, Monitor, UserStatsValue};
+use crate::statistics::{user_stats::UserStatsValue, ClientStats};
+use crate::{monitors::Monitor, statistics::manager::ClientStatsManager};
 
 /// Prometheus metrics for global and each client.
 #[derive(Clone, Debug, Default)]
@@ -73,7 +73,6 @@ where
     prometheus_global_stats: PrometheusStats, // global prometheus metrics
     prometheus_client_stats: PrometheusStats, // per-client prometheus metrics
     client_stats: Vec<ClientStats>,           // per-client statistics
-    aggregator: Aggregator,                   // aggregator for global custom statistics
 }
 
 impl<F> Debug for PrometheusMonitor<F>
@@ -92,16 +91,6 @@ impl<F> Monitor for PrometheusMonitor<F>
 where
     F: FnMut(&str),
 {
-    /// the client monitor, mutable
-    fn client_stats_mut(&mut self) -> &mut Vec<ClientStats> {
-        &mut self.client_stats
-    }
-
-    /// the client monitor
-    fn client_stats(&self) -> &[ClientStats] {
-        &self.client_stats
-    }
-
     /// Time this fuzzing run stated
     fn start_time(&self) -> Duration {
         self.start_time
@@ -112,12 +101,12 @@ where
         self.start_time = time;
     }
 
-    /// aggregate client stats
-    fn aggregate(&mut self, name: &str) {
-        self.aggregator.aggregate(name, &self.client_stats);
-    }
-
-    fn display(&mut self, event_msg: &str, sender_id: ClientId) {
+    fn display(
+        &mut self,
+        client_stats_manager: &mut ClientStatsManager,
+        event_msg: &str,
+        sender_id: ClientId,
+    ) {
         // Update the prometheus metrics
         // The gauges must take signed i64's, with max value of 2^63-1 so it is
         // probably fair to error out at a count of nine quintillion across any
@@ -127,7 +116,7 @@ where
         // time since last observation"
 
         // Global (aggregated) metrics
-        let corpus_size = self.corpus_size();
+        let corpus_size = client_stats_manager.corpus_size();
         self.prometheus_global_stats
             .corpus_count
             .get_or_create(&Labels {
@@ -136,7 +125,7 @@ where
             })
             .set(corpus_size.try_into().unwrap());
 
-        let objective_size = self.objective_size();
+        let objective_size = client_stats_manager.objective_size();
         self.prometheus_global_stats
             .objective_count
             .get_or_create(&Labels {
@@ -145,7 +134,7 @@ where
             })
             .set(objective_size.try_into().unwrap());
 
-        let total_execs = self.total_execs();
+        let total_execs = client_stats_manager.total_execs();
         self.prometheus_global_stats
             .executions
             .get_or_create(&Labels {
@@ -154,7 +143,7 @@ where
             })
             .set(total_execs.try_into().unwrap());
 
-        let execs_per_sec = self.execs_per_sec();
+        let execs_per_sec = client_stats_manager.execs_per_sec();
         self.prometheus_global_stats
             .exec_rate
             .get_or_create(&Labels {
@@ -172,7 +161,10 @@ where
             })
             .set(run_time.try_into().unwrap()); // run time in seconds, which can be converted to a time format by Grafana or similar
 
-        let total_clients = self.client_stats_count().try_into().unwrap(); // convert usize to u64 (unlikely that # of clients will be > 2^64 -1...)
+        let total_clients = client_stats_manager
+            .client_stats_count()
+            .try_into()
+            .unwrap(); // convert usize to u64 (unlikely that # of clients will be > 2^64 -1...)
         self.prometheus_global_stats
             .clients_count
             .get_or_create(&Labels {
@@ -186,13 +178,13 @@ where
             "[Prometheus] [{} #GLOBAL] run time: {}, clients: {}, corpus: {}, objectives: {}, executions: {}, exec/sec: {}",
             event_msg,
             format_duration_hms(&(current_time() - self.start_time)),
-            self.client_stats_count(),
-            self.corpus_size(),
-            self.objective_size(),
-            self.total_execs(),
-            self.execs_per_sec_pretty()
+            client_stats_manager.client_stats_count(),
+            client_stats_manager.corpus_size(),
+            client_stats_manager.objective_size(),
+            client_stats_manager.total_execs(),
+            client_stats_manager.execs_per_sec_pretty()
         );
-        for (key, val) in &self.aggregator.aggregated {
+        for (key, val) in client_stats_manager.aggregated() {
             // print global aggregated custom stats
             write!(global_fmt, ", {key}: {val}").unwrap();
             #[expect(clippy::cast_precision_loss)]
@@ -234,8 +226,8 @@ where
 
         // Client-specific metrics
 
-        self.client_stats_insert(sender_id);
-        let client = self.client_stats_for(sender_id);
+        client_stats_manager.client_stats_insert(sender_id);
+        let client = client_stats_manager.client_stats_for(sender_id);
         let mut cur_client_clone = client.clone();
 
         self.prometheus_client_stats
@@ -297,7 +289,7 @@ where
             cur_client_clone.execs_per_sec_pretty(current_time())
         );
 
-        for (key, val) in cur_client_clone.user_monitor {
+        for (key, val) in cur_client_clone.user_stats {
             // print the custom stats for each client
             write!(fmt, ", {key}: {val}").unwrap();
             // Update metrics added to the user_stats hashmap by feedback event-fires
@@ -370,7 +362,6 @@ where
             prometheus_global_stats,
             prometheus_client_stats,
             client_stats,
-            aggregator: Aggregator::new(),
         }
     }
     /// Creates the monitor with a given `start_time`.
@@ -396,7 +387,6 @@ where
             prometheus_global_stats,
             prometheus_client_stats,
             client_stats,
-            aggregator: Aggregator::new(),
         }
     }
 }
